refactor(routes): give the route table a module without components
web/routes imports App, so every component reading a path from it forms
a cycle. That already cost us once: reading the map at module level in
App got an empty object and the app failed to boot, worked around by
reading it lazily inside the function.

The table moves to web/routePaths, which imports nothing but
react-router. web/routes re-exports it, so nothing outside changes, and
the three components that wanted a path now take it from there. The
lazy read goes with the cycle.

isAllowedWhileBlocked moves too, and is now testable without pulling in
the app. jest gains the web/ alias webpack already resolves, without
which it cannot follow those imports.
